refactor(client): extract params and response reading into submodules
Move parameter conversion (convert_params, encode_tvp_param,
convert_tvp_column_type, encode_tvp_value) into client/params.rs and
response reading (read_query_response, read_execute_result,
read_transaction_begin_result, read_multi_result_response) into
client/response.rs.

client/mod.rs: 1,930 → 1,047 lines (-46%)
New client/params.rs: 353 lines
New client/response.rs: 565 lines

Zero public API changes. Zero warnings across all feature configurations.
