+++ crates/mssql-client/src/client/mod.rs
@@ -5,22 +5,17 @@
 #![allow(clippy::unwrap_used, clippy::expect_used, clippy::needless_range_loop)]
 
 mod connect;
+mod params;
+mod response;
 
 use std::marker::PhantomData;
 
-use bytes::BytesMut;
 use mssql_codec::connection::Connection;
 #[cfg(feature = "tls")]
 use mssql_tls::TlsStream;
 use tds_protocol::packet::PacketType;
-use tds_protocol::rpc::{RpcParam, RpcRequest, TypeInfo as RpcTypeInfo};
-use tds_protocol::token::{ColMetaData, EnvChange, EnvChangeType, Token, TokenParser};
-#[cfg(feature = "decimal")]
-use tds_protocol::tvp::encode_tvp_decimal;
-use tds_protocol::tvp::{
-    TvpColumnDef as TvpWireColumnDef, TvpColumnFlags, TvpEncoder, TvpWireType, encode_tvp_bit,
-    encode_tvp_float, encode_tvp_int, encode_tvp_null, encode_tvp_nvarchar, encode_tvp_varbinary,
-};
+use tds_protocol::rpc::RpcRequest;
+use tds_protocol::token::{EnvChange, EnvChangeType};
 use tokio::net::TcpStream;
 use tokio::time::timeout;
 
@@ -201,700 +196,6 @@ impl<S: ConnectionState> Client<S> {
 
         Ok(())
     }
-
-    /// Convert ToSql parameters to RPC parameters.
-    fn convert_params(params: &[&(dyn crate::ToSql + Sync)]) -> Result<Vec<RpcParam>> {
-        use bytes::{BufMut, BytesMut};
-        use mssql_types::SqlValue;
-
-        params
-            .iter()
-            .enumerate()
-            .map(|(i, p)| {
-                let sql_value = p.to_sql()?;
-                let name = format!("@p{}", i + 1);
-
-                Ok(match sql_value {
-                    SqlValue::Null => RpcParam::null(&name, RpcTypeInfo::nvarchar(1)),
-                    SqlValue::Bool(v) => {
-                        let mut buf = BytesMut::with_capacity(1);
-                        buf.put_u8(if v { 1 } else { 0 });
-                        RpcParam::new(&name, RpcTypeInfo::bit(), buf.freeze())
-                    }
-                    SqlValue::TinyInt(v) => {
-                        let mut buf = BytesMut::with_capacity(1);
-                        buf.put_u8(v);
-                        RpcParam::new(&name, RpcTypeInfo::tinyint(), buf.freeze())
-                    }
-                    SqlValue::SmallInt(v) => {
-                        let mut buf = BytesMut::with_capacity(2);
-                        buf.put_i16_le(v);
-                        RpcParam::new(&name, RpcTypeInfo::smallint(), buf.freeze())
-                    }
-                    SqlValue::Int(v) => RpcParam::int(&name, v),
-                    SqlValue::BigInt(v) => RpcParam::bigint(&name, v),
-                    SqlValue::Float(v) => {
-                        let mut buf = BytesMut::with_capacity(4);
-                        buf.put_f32_le(v);
-                        RpcParam::new(&name, RpcTypeInfo::real(), buf.freeze())
-                    }
-                    SqlValue::Double(v) => {
-                        let mut buf = BytesMut::with_capacity(8);
-                        buf.put_f64_le(v);
-                        RpcParam::new(&name, RpcTypeInfo::float(), buf.freeze())
-                    }
-                    SqlValue::String(ref s) => RpcParam::nvarchar(&name, s),
-                    SqlValue::Binary(ref b) => {
-                        RpcParam::new(&name, RpcTypeInfo::varbinary(b.len() as u16), b.clone())
-                    }
-                    SqlValue::Xml(ref s) => RpcParam::nvarchar(&name, s),
-                    #[cfg(feature = "uuid")]
-                    SqlValue::Uuid(u) => {
-                        // UUID is stored in a specific byte order for SQL Server
-                        let bytes = u.as_bytes();
-                        let mut buf = BytesMut::with_capacity(16);
-                        // SQL Server stores GUIDs in mixed-endian format
-                        buf.put_u32_le(u32::from_be_bytes([
-                            bytes[0], bytes[1], bytes[2], bytes[3],
-                        ]));
-                        buf.put_u16_le(u16::from_be_bytes([bytes[4], bytes[5]]));
-                        buf.put_u16_le(u16::from_be_bytes([bytes[6], bytes[7]]));
-                        buf.put_slice(&bytes[8..16]);
-                        RpcParam::new(&name, RpcTypeInfo::uniqueidentifier(), buf.freeze())
-                    }
-                    #[cfg(feature = "decimal")]
-                    SqlValue::Decimal(d) => {
-                        // Decimal encoding is complex; use string representation for now
-                        RpcParam::nvarchar(&name, &d.to_string())
-                    }
-                    #[cfg(feature = "chrono")]
-                    SqlValue::Date(_)
-                    | SqlValue::Time(_)
-                    | SqlValue::DateTime(_)
-                    | SqlValue::DateTimeOffset(_) => {
-                        // For date/time types, use string representation for simplicity
-                        // A full implementation would encode these properly
-                        let s = match &sql_value {
-                            SqlValue::Date(d) => d.to_string(),
-                            SqlValue::Time(t) => t.to_string(),
-                            SqlValue::DateTime(dt) => dt.to_string(),
-                            SqlValue::DateTimeOffset(dto) => dto.to_rfc3339(),
-                            _ => unreachable!(),
-                        };
-                        RpcParam::nvarchar(&name, &s)
-                    }
-                    #[cfg(feature = "json")]
-                    SqlValue::Json(ref j) => RpcParam::nvarchar(&name, &j.to_string()),
-                    SqlValue::Tvp(ref tvp_data) => {
-                        // Encode TVP using the wire format
-                        Self::encode_tvp_param(&name, tvp_data)?
-                    }
-                    // Handle future SqlValue variants
-                    _ => {
-                        return Err(Error::Type(mssql_types::TypeError::UnsupportedConversion {
-                            from: sql_value.type_name().to_string(),
-                            to: "RPC parameter",
-                        }));
-                    }
-                })
-            })
-            .collect()
-    }
-
-    /// Encode a TVP parameter for RPC.
-    ///
-    /// This encodes the complete TVP structure including metadata and row data
-    /// into the TDS wire format.
-    fn encode_tvp_param(name: &str, tvp_data: &mssql_types::TvpData) -> Result<RpcParam> {
-        // Convert mssql-types column definitions to wire format
-        let wire_columns: Vec<TvpWireColumnDef> = tvp_data
-            .columns
-            .iter()
-            .map(|col| {
-                let wire_type = Self::convert_tvp_column_type(&col.column_type);
-                TvpWireColumnDef {
-                    wire_type,
-                    flags: TvpColumnFlags {
-                        nullable: col.nullable,
-                    },
-                }
-            })
-            .collect();
-
-        // Create encoder
-        let encoder = TvpEncoder::new(&tvp_data.schema, &tvp_data.type_name, &wire_columns);
-
-        // Encode to buffer
-        let mut buf = BytesMut::with_capacity(256);
-
-        // Encode metadata
-        encoder.encode_metadata(&mut buf);
-
-        // Encode each row
-        for row in &tvp_data.rows {
-            encoder.encode_row(&mut buf, |row_buf| {
-                for (col_idx, value) in row.iter().enumerate() {
-                    let wire_type = &wire_columns[col_idx].wire_type;
-                    Self::encode_tvp_value(value, wire_type, row_buf);
-                }
-            });
-        }
-
-        // Encode end marker
-        encoder.encode_end(&mut buf);
-
-        // Build the full TVP type name (schema.TypeName)
-        let full_type_name = if tvp_data.schema.is_empty() {
-            tvp_data.type_name.clone()
-        } else {
-            format!("{}.{}", tvp_data.schema, tvp_data.type_name)
-        };
-
-        // Create RPC param with TVP type info
-        // The type info includes the TVP type name for parameter declarations
-        let type_info = RpcTypeInfo::tvp(&full_type_name);
-
-        Ok(RpcParam {
-            name: name.to_string(),
-            flags: tds_protocol::rpc::ParamFlags::default(),
-            type_info,
-            value: Some(buf.freeze()),
-        })
-    }
-
-    /// Convert mssql-types TvpColumnType to wire TvpWireType.
-    fn convert_tvp_column_type(col_type: &mssql_types::TvpColumnType) -> TvpWireType {
-        match col_type {
-            mssql_types::TvpColumnType::Bit => TvpWireType::Bit,
-            mssql_types::TvpColumnType::TinyInt => TvpWireType::Int { size: 1 },
-            mssql_types::TvpColumnType::SmallInt => TvpWireType::Int { size: 2 },
-            mssql_types::TvpColumnType::Int => TvpWireType::Int { size: 4 },
-            mssql_types::TvpColumnType::BigInt => TvpWireType::Int { size: 8 },
-            mssql_types::TvpColumnType::Real => TvpWireType::Float { size: 4 },
-            mssql_types::TvpColumnType::Float => TvpWireType::Float { size: 8 },
-            mssql_types::TvpColumnType::Decimal { precision, scale } => TvpWireType::Decimal {
-                precision: *precision,
-                scale: *scale,
-            },
-            mssql_types::TvpColumnType::NVarChar { max_length } => TvpWireType::NVarChar {
-                max_length: *max_length,
-            },
-            mssql_types::TvpColumnType::VarChar { max_length } => TvpWireType::VarChar {
-                max_length: *max_length,
-            },
-            mssql_types::TvpColumnType::VarBinary { max_length } => TvpWireType::VarBinary {
-                max_length: *max_length,
-            },
-            mssql_types::TvpColumnType::UniqueIdentifier => TvpWireType::Guid,
-            mssql_types::TvpColumnType::Date => TvpWireType::Date,
-            mssql_types::TvpColumnType::Time { scale } => TvpWireType::Time { scale: *scale },
-            mssql_types::TvpColumnType::DateTime2 { scale } => {
-                TvpWireType::DateTime2 { scale: *scale }
-            }
-            mssql_types::TvpColumnType::DateTimeOffset { scale } => {
-                TvpWireType::DateTimeOffset { scale: *scale }
-            }
-            mssql_types::TvpColumnType::Xml => TvpWireType::Xml,
-        }
-    }
-
-    /// Encode a single TVP column value.
-    fn encode_tvp_value(
-        value: &mssql_types::SqlValue,
-        wire_type: &TvpWireType,
-        buf: &mut BytesMut,
-    ) {
-        use mssql_types::SqlValue;
-
-        match value {
-            SqlValue::Null => {
-                encode_tvp_null(wire_type, buf);
-            }
-            SqlValue::Bool(v) => {
-                encode_tvp_bit(*v, buf);
-            }
-            SqlValue::TinyInt(v) => {
-                encode_tvp_int(*v as i64, 1, buf);
-            }
-            SqlValue::SmallInt(v) => {
-                encode_tvp_int(*v as i64, 2, buf);
-            }
-            SqlValue::Int(v) => {
-                encode_tvp_int(*v as i64, 4, buf);
-            }
-            SqlValue::BigInt(v) => {
-                encode_tvp_int(*v, 8, buf);
-            }
-            SqlValue::Float(v) => {
-                encode_tvp_float(*v as f64, 4, buf);
-            }
-            SqlValue::Double(v) => {
-                encode_tvp_float(*v, 8, buf);
-            }
-            SqlValue::String(s) => {
-                let max_len = match wire_type {
-                    TvpWireType::NVarChar { max_length } => *max_length,
-                    _ => 4000,
-                };
-                encode_tvp_nvarchar(s, max_len, buf);
-            }
-            SqlValue::Binary(b) => {
-                let max_len = match wire_type {
-                    TvpWireType::VarBinary { max_length } => *max_length,
-                    _ => 8000,
-                };
-                encode_tvp_varbinary(b, max_len, buf);
-            }
-            #[cfg(feature = "decimal")]
-            SqlValue::Decimal(d) => {
-                let sign = if d.is_sign_negative() { 0u8 } else { 1u8 };
-                let mantissa = d.mantissa().unsigned_abs();
-                encode_tvp_decimal(sign, mantissa, buf);
-            }
-            #[cfg(feature = "uuid")]
-            SqlValue::Uuid(u) => {
-                let bytes = u.as_bytes();
-                tds_protocol::tvp::encode_tvp_guid(bytes, buf);
-            }
-            #[cfg(feature = "chrono")]
-            SqlValue::Date(d) => {
-                // Calculate days since 0001-01-01
-                let base = chrono::NaiveDate::from_ymd_opt(1, 1, 1).unwrap();
-                let days = d.signed_duration_since(base).num_days() as u32;
-                tds_protocol::tvp::encode_tvp_date(days, buf);
-            }
-            #[cfg(feature = "chrono")]
-            SqlValue::Time(t) => {
-                use chrono::Timelike;
-                let nanos =
-                    t.num_seconds_from_midnight() as u64 * 1_000_000_000 + t.nanosecond() as u64;
-                let intervals = nanos / 100;
-                let scale = match wire_type {
-                    TvpWireType::Time { scale } => *scale,
-                    _ => 7,
-                };
-                tds_protocol::tvp::encode_tvp_time(intervals, scale, buf);
-            }
-            #[cfg(feature = "chrono")]
-            SqlValue::DateTime(dt) => {
-                use chrono::Timelike;
-                // Time component
-                let nanos = dt.time().num_seconds_from_midnight() as u64 * 1_000_000_000
-                    + dt.time().nanosecond() as u64;
-                let intervals = nanos / 100;
-                // Date component
-                let base = chrono::NaiveDate::from_ymd_opt(1, 1, 1).unwrap();
-                let days = dt.date().signed_duration_since(base).num_days() as u32;
-                let scale = match wire_type {
-                    TvpWireType::DateTime2 { scale } => *scale,
-                    _ => 7,
-                };
-                tds_protocol::tvp::encode_tvp_datetime2(intervals, days, scale, buf);
-            }
-            #[cfg(feature = "chrono")]
-            SqlValue::DateTimeOffset(dto) => {
-                use chrono::{Offset, Timelike};
-                // Time component (in 100-nanosecond intervals)
-                let nanos = dto.time().num_seconds_from_midnight() as u64 * 1_000_000_000
-                    + dto.time().nanosecond() as u64;
-                let intervals = nanos / 100;
-                // Date component (days since 0001-01-01)
-                let base = chrono::NaiveDate::from_ymd_opt(1, 1, 1).unwrap();
-                let days = dto.date_naive().signed_duration_since(base).num_days() as u32;
-                // Timezone offset in minutes
-                let offset_minutes = (dto.offset().fix().local_minus_utc() / 60) as i16;
-                let scale = match wire_type {
-                    TvpWireType::DateTimeOffset { scale } => *scale,
-                    _ => 7,
-                };
-                tds_protocol::tvp::encode_tvp_datetimeoffset(
-                    intervals,
-                    days,
-                    offset_minutes,
-                    scale,
-                    buf,
-                );
-            }
-            #[cfg(feature = "json")]
-            SqlValue::Json(j) => {
-                // JSON is encoded as NVARCHAR
-                encode_tvp_nvarchar(&j.to_string(), 0xFFFF, buf);
-            }
-            SqlValue::Xml(s) => {
-                // XML is encoded as NVARCHAR for TVP
-                encode_tvp_nvarchar(s, 0xFFFF, buf);
-            }
-            SqlValue::Tvp(_) => {
-                // Nested TVPs are not supported
-                encode_tvp_null(wire_type, buf);
-            }
-            // Handle future SqlValue variants as NULL
-            _ => {
-                encode_tvp_null(wire_type, buf);
-            }
-        }
-    }
-
-    /// Read complete query response including columns and rows.
-    async fn read_query_response(
-        &mut self,
-    ) -> Result<(Vec<crate::row::Column>, Vec<crate::row::Row>)> {
-        let connection = self.connection.as_mut().ok_or(Error::ConnectionClosed)?;
-
-        let message = match connection {
-            #[cfg(feature = "tls")]
-            ConnectionHandle::Tls(conn) => conn
-                .read_message()
-                .await
-                .map_err(|e| Error::Protocol(e.to_string()))?,
-            #[cfg(feature = "tls")]
-            ConnectionHandle::TlsPrelogin(conn) => conn
-                .read_message()
-                .await
-                .map_err(|e| Error::Protocol(e.to_string()))?,
-            ConnectionHandle::Plain(conn) => conn
-                .read_message()
-                .await
-                .map_err(|e| Error::Protocol(e.to_string()))?,
-        }
-        .ok_or(Error::ConnectionClosed)?;
-
-        let mut parser = TokenParser::new(message.payload);
-        let mut columns: Vec<crate::row::Column> = Vec::new();
-        let mut rows: Vec<crate::row::Row> = Vec::new();
-        let mut protocol_metadata: Option<ColMetaData> = None;
-
-        loop {
-            // Use next_token_with_metadata to properly parse Row/NbcRow tokens
-            let token = parser
-                .next_token_with_metadata(protocol_metadata.as_ref())
-                .map_err(|e| Error::Protocol(e.to_string()))?;
-
-            let Some(token) = token else {
-                break;
-            };
-
-            match token {
-                Token::ColMetaData(meta) => {
-                    // New result set starting - clear previous rows
-                    // This enables multi-statement batches to return the last result set
-                    rows.clear();
-
-                    columns = meta
-                        .columns
-                        .iter()
-                        .enumerate()
-                        .map(|(i, col)| {
-                            let type_name = format!("{:?}", col.type_id);
-                            let mut column = crate::row::Column::new(&col.name, i, type_name)
-                                .with_nullable(col.flags & 0x01 != 0);
-
-                            if let Some(max_len) = col.type_info.max_length {
-                                column = column.with_max_length(max_len);
-                            }
-                            if let (Some(prec), Some(scale)) =
-                                (col.type_info.precision, col.type_info.scale)
-                            {
-                                column = column.with_precision_scale(prec, scale);
-                            }
-                            // Store collation for VARCHAR/CHAR types to enable
-                            // collation-aware string decoding
-                            if let Some(collation) = col.type_info.collation {
-                                column = column.with_collation(collation);
-                            }
-                            column
-                        })
-                        .collect();
-
-                    tracing::debug!(columns = columns.len(), "received column metadata");
-                    protocol_metadata = Some(meta);
-                }
-                Token::Row(raw_row) => {
-                    if let Some(ref meta) = protocol_metadata {
-                        let row = crate::column_parser::convert_raw_row(&raw_row, meta, &columns)?;
-                        rows.push(row);
-                    }
-                }
-                Token::NbcRow(nbc_row) => {
-                    if let Some(ref meta) = protocol_metadata {
-                        let row = crate::column_parser::convert_nbc_row(&nbc_row, meta, &columns)?;
-                        rows.push(row);
-                    }
-                }
-                Token::Error(err) => {
-                    return Err(Error::Server {
-                        number: err.number,
-                        state: err.state,
-                        class: err.class,
-                        message: err.message.clone(),
-                        server: if err.server.is_empty() {
-                            None
-                        } else {
-                            Some(err.server.clone())
-                        },
-                        procedure: if err.procedure.is_empty() {
-                            None
-                        } else {
-                            Some(err.procedure.clone())
-                        },
-                        line: err.line as u32,
-                    });
-                }
-                Token::Done(done) => {
-                    if done.status.error {
-                        return Err(Error::Query("query failed".to_string()));
-                    }
-                    tracing::debug!(
-                        row_count = done.row_count,
-                        has_more = done.status.more,
-                        "query complete"
-                    );
-                    // Only break if there are no more result sets
-                    // This enables multi-statement batches to process all results
-                    if !done.status.more {
-                        break;
-                    }
-                }
-                Token::DoneProc(done) => {
-                    if done.status.error {
-                        return Err(Error::Query("query failed".to_string()));
-                    }
-                }
-                Token::DoneInProc(done) => {
-                    if done.status.error {
-                        return Err(Error::Query("query failed".to_string()));
-                    }
-                }
-                Token::Info(info) => {
-                    tracing::debug!(
-                        number = info.number,
-                        message = %info.message,
-                        "server info message"
-                    );
-                }
-                Token::EnvChange(env) => {
-                    // Process transaction-related EnvChange tokens.
-                    // This allows BEGIN TRANSACTION, COMMIT, ROLLBACK via raw SQL
-                    // to properly update the transaction descriptor.
-                    Self::process_transaction_env_change(&env, &mut self.transaction_descriptor);
-                }
-                _ => {}
-            }
-        }
-
-        tracing::debug!(
-            columns = columns.len(),
-            rows = rows.len(),
-            "query response parsed"
-        );
-        Ok((columns, rows))
-    }
-
-    /// Read execute result (row count) from the response.
-    async fn read_execute_result(&mut self) -> Result<u64> {
-        let connection = self.connection.as_mut().ok_or(Error::ConnectionClosed)?;
-
-        let message = match connection {
-            #[cfg(feature = "tls")]
-            ConnectionHandle::Tls(conn) => conn
-                .read_message()
-                .await
-                .map_err(|e| Error::Protocol(e.to_string()))?,
-            #[cfg(feature = "tls")]
-            ConnectionHandle::TlsPrelogin(conn) => conn
-                .read_message()
-                .await
-                .map_err(|e| Error::Protocol(e.to_string()))?,
-            ConnectionHandle::Plain(conn) => conn
-                .read_message()
-                .await
-                .map_err(|e| Error::Protocol(e.to_string()))?,
-        }
-        .ok_or(Error::ConnectionClosed)?;
-
-        let mut parser = TokenParser::new(message.payload);
-        let mut rows_affected = 0u64;
-        let mut current_metadata: Option<ColMetaData> = None;
-
-        loop {
-            // Use metadata-aware parsing to handle Row tokens from SELECT statements
-            let token = parser
-                .next_token_with_metadata(current_metadata.as_ref())
-                .map_err(|e| Error::Protocol(e.to_string()))?;
-
-            let Some(token) = token else {
-                break;
-            };
-
-            match token {
-                Token::ColMetaData(meta) => {
-                    // Store metadata for subsequent Row token parsing
-                    current_metadata = Some(meta);
-                }
-                Token::Row(_) | Token::NbcRow(_) => {
-                    // Skip row data for execute() - we only care about row count
-                    // The rows are parsed but we don't process them
-                }
-                Token::Done(done) => {
-                    if done.status.error {
-                        return Err(Error::Query("execution failed".to_string()));
-                    }
-                    if done.status.count {
-                        // Accumulate row counts from all statements in a batch
-                        rows_affected += done.row_count;
-                    }
-                    // Only break if there are no more result sets
-                    // This enables multi-statement batches to report total affected rows
-                    if !done.status.more {
-                        break;
-                    }
-                }
-                Token::DoneProc(done) => {
-                    if done.status.count {
-                        rows_affected += done.row_count;
-                    }
-                }
-                Token::DoneInProc(done) => {
-                    if done.status.count {
-                        rows_affected += done.row_count;
-                    }
-                }
-                Token::Error(err) => {
-                    return Err(Error::Server {
-                        number: err.number,
-                        state: err.state,
-                        class: err.class,
-                        message: err.message.clone(),
-                        server: if err.server.is_empty() {
-                            None
-                        } else {
-                            Some(err.server.clone())
-                        },
-                        procedure: if err.procedure.is_empty() {
-                            None
-                        } else {
-                            Some(err.procedure.clone())
-                        },
-                        line: err.line as u32,
-                    });
-                }
-                Token::Info(info) => {
-                    tracing::info!(
-                        number = info.number,
-                        message = %info.message,
-                        "server info message"
-                    );
-                }
-                Token::EnvChange(env) => {
-                    // Process transaction-related EnvChange tokens.
-                    // This allows BEGIN TRANSACTION, COMMIT, ROLLBACK via raw SQL
-                    // to properly update the transaction descriptor.
-                    Self::process_transaction_env_change(&env, &mut self.transaction_descriptor);
-                }
-                _ => {}
-            }
-        }
-
-        Ok(rows_affected)
-    }
-
-    /// Read the response from BEGIN TRANSACTION and extract the transaction descriptor.
-    ///
-    /// Per MS-TDS spec, the server sends a BeginTransaction EnvChange token containing
-    /// the transaction descriptor (8-byte value) that must be included in subsequent
-    /// ALL_HEADERS sections for requests within this transaction.
-    async fn read_transaction_begin_result(&mut self) -> Result<u64> {
-        let connection = self.connection.as_mut().ok_or(Error::ConnectionClosed)?;
-
-        let message = match connection {
-            #[cfg(feature = "tls")]
-            ConnectionHandle::Tls(conn) => conn
-                .read_message()
-                .await
-                .map_err(|e| Error::Protocol(e.to_string()))?,
-            #[cfg(feature = "tls")]
-            ConnectionHandle::TlsPrelogin(conn) => conn
-                .read_message()
-                .await
-                .map_err(|e| Error::Protocol(e.to_string()))?,
-            ConnectionHandle::Plain(conn) => conn
-                .read_message()
-                .await
-                .map_err(|e| Error::Protocol(e.to_string()))?,
-        }
-        .ok_or(Error::ConnectionClosed)?;
-
-        let mut parser = TokenParser::new(message.payload);
-        let mut transaction_descriptor: u64 = 0;
-
-        loop {
-            let token = parser
-                .next_token()
-                .map_err(|e| Error::Protocol(e.to_string()))?;
-
-            let Some(token) = token else {
-                break;
-            };
-
-            match token {
-                Token::EnvChange(env) => {
-                    if env.env_type == EnvChangeType::BeginTransaction {
-                        // Extract the transaction descriptor from the binary value
-                        // Per MS-TDS spec, it's an 8-byte (ULONGLONG) value
-                        if let tds_protocol::token::EnvChangeValue::Binary(ref data) = env.new_value
-                        {
-                            if data.len() >= 8 {
-                                transaction_descriptor = u64::from_le_bytes([
-                                    data[0], data[1], data[2], data[3], data[4], data[5], data[6],
-                                    data[7],
-                                ]);
-                                tracing::debug!(
-                                    transaction_descriptor =
-                                        format!("0x{:016X}", transaction_descriptor),
-                                    "transaction begun"
-                                );
-                            }
-                        }
-                    }
-                }
-                Token::Done(done) => {
-                    if done.status.error {
-                        return Err(Error::Query("BEGIN TRANSACTION failed".to_string()));
-                    }
-                    break;
-                }
-                Token::Error(err) => {
-                    return Err(Error::Server {
-                        number: err.number,
-                        state: err.state,
-                        class: err.class,
-                        message: err.message.clone(),
-                        server: if err.server.is_empty() {
-                            None
-                        } else {
-                            Some(err.server.clone())
-                        },
-                        procedure: if err.procedure.is_empty() {
-                            None
-                        } else {
-                            Some(err.procedure.clone())
-                        },
-                        line: err.line as u32,
-                    });
-                }
-                Token::Info(info) => {
-                    tracing::info!(
-                        number = info.number,
-                        message = %info.message,
-                        "server info message"
-                    );
-                }
-                _ => {}
-            }
-        }
-
-        Ok(transaction_descriptor)
-    }
 }
 
 impl Client<Ready> {
@@ -1074,190 +375,6 @@ impl Client<Ready> {
         // Read all result sets
         let result_sets = self.read_multi_result_response().await?;
         Ok(MultiResultStream::new(result_sets))
-    }
-
-    /// Read multiple result sets from a batch response.
-    async fn read_multi_result_response(&mut self) -> Result<Vec<crate::stream::ResultSet>> {
-        let connection = self.connection.as_mut().ok_or(Error::ConnectionClosed)?;
-
-        let message = match connection {
-            #[cfg(feature = "tls")]
-            ConnectionHandle::Tls(conn) => conn
-                .read_message()
-                .await
-                .map_err(|e| Error::Protocol(e.to_string()))?,
-            #[cfg(feature = "tls")]
-            ConnectionHandle::TlsPrelogin(conn) => conn
-                .read_message()
-                .await
-                .map_err(|e| Error::Protocol(e.to_string()))?,
-            ConnectionHandle::Plain(conn) => conn
-                .read_message()
-                .await
-                .map_err(|e| Error::Protocol(e.to_string()))?,
-        }
-        .ok_or(Error::ConnectionClosed)?;
-
-        let mut parser = TokenParser::new(message.payload);
-        let mut result_sets: Vec<crate::stream::ResultSet> = Vec::new();
-        let mut current_columns: Vec<crate::row::Column> = Vec::new();
-        let mut current_rows: Vec<crate::row::Row> = Vec::new();
-        let mut protocol_metadata: Option<ColMetaData> = None;
-
-        loop {
-            let token = parser
-                .next_token_with_metadata(protocol_metadata.as_ref())
-                .map_err(|e| Error::Protocol(e.to_string()))?;
-
-            let Some(token) = token else {
-                break;
-            };
-
-            match token {
-                Token::ColMetaData(meta) => {
-                    // New result set starting - save the previous one if it has columns
-                    if !current_columns.is_empty() {
-                        result_sets.push(crate::stream::ResultSet::new(
-                            std::mem::take(&mut current_columns),
-                            std::mem::take(&mut current_rows),
-                        ));
-                    }
-
-                    // Parse the new column metadata
-                    current_columns = meta
-                        .columns
-                        .iter()
-                        .enumerate()
-                        .map(|(i, col)| {
-                            let type_name = format!("{:?}", col.type_id);
-                            let mut column = crate::row::Column::new(&col.name, i, type_name)
-                                .with_nullable(col.flags & 0x01 != 0);
-
-                            if let Some(max_len) = col.type_info.max_length {
-                                column = column.with_max_length(max_len);
-                            }
-                            if let (Some(prec), Some(scale)) =
-                                (col.type_info.precision, col.type_info.scale)
-                            {
-                                column = column.with_precision_scale(prec, scale);
-                            }
-                            // Store collation for VARCHAR/CHAR types to enable
-                            // collation-aware string decoding
-                            if let Some(collation) = col.type_info.collation {
-                                column = column.with_collation(collation);
-                            }
-                            column
-                        })
-                        .collect();
-
-                    tracing::debug!(
-                        columns = current_columns.len(),
-                        result_set = result_sets.len(),
-                        "received column metadata for result set"
-                    );
-                    protocol_metadata = Some(meta);
-                }
-                Token::Row(raw_row) => {
-                    if let Some(ref meta) = protocol_metadata {
-                        let row = crate::column_parser::convert_raw_row(
-                            &raw_row,
-                            meta,
-                            &current_columns,
-                        )?;
-                        current_rows.push(row);
-                    }
-                }
-                Token::NbcRow(nbc_row) => {
-                    if let Some(ref meta) = protocol_metadata {
-                        let row = crate::column_parser::convert_nbc_row(
-                            &nbc_row,
-                            meta,
-                            &current_columns,
-                        )?;
-                        current_rows.push(row);
-                    }
-                }
-                Token::Error(err) => {
-                    return Err(Error::Server {
-                        number: err.number,
-                        state: err.state,
-                        class: err.class,
-                        message: err.message.clone(),
-                        server: if err.server.is_empty() {
-                            None
-                        } else {
-                            Some(err.server.clone())
-                        },
-                        procedure: if err.procedure.is_empty() {
-                            None
-                        } else {
-                            Some(err.procedure.clone())
-                        },
-                        line: err.line as u32,
-                    });
-                }
-                Token::Done(done) => {
-                    if done.status.error {
-                        return Err(Error::Query("query failed".to_string()));
-                    }
-
-                    // Save the current result set if we have columns
-                    if !current_columns.is_empty() {
-                        result_sets.push(crate::stream::ResultSet::new(
-                            std::mem::take(&mut current_columns),
-                            std::mem::take(&mut current_rows),
-                        ));
-                        protocol_metadata = None;
-                    }
-
-                    // Check if there are more result sets
-                    if !done.status.more {
-                        tracing::debug!(result_sets = result_sets.len(), "all result sets parsed");
-                        break;
-                    }
-                }
-                Token::DoneInProc(done) => {
-                    if done.status.error {
-                        return Err(Error::Query("query failed".to_string()));
-                    }
-
-                    // Save the current result set if we have columns (within stored proc)
-                    if !current_columns.is_empty() {
-                        result_sets.push(crate::stream::ResultSet::new(
-                            std::mem::take(&mut current_columns),
-                            std::mem::take(&mut current_rows),
-                        ));
-                        protocol_metadata = None;
-                    }
-
-                    // DoneInProc may indicate more results within the batch
-                    if !done.status.more {
-                        // No more results from this statement, but batch may continue
-                    }
-                }
-                Token::DoneProc(done) => {
-                    if done.status.error {
-                        return Err(Error::Query("query failed".to_string()));
-                    }
-                    // DoneProc marks end of stored procedure, not necessarily end of results
-                }
-                Token::Info(info) => {
-                    tracing::debug!(
-                        number = info.number,
-                        message = %info.message,
-                        "server info message"
-                    );
-                }
-                _ => {}
-            }
-        }
-
-        // Don't forget any remaining result set that wasn't followed by Done
-        if !current_columns.is_empty() {
-            result_sets.push(crate::stream::ResultSet::new(current_columns, current_rows));
-        }
-
-        Ok(result_sets)
     }
 
     /// Execute a query that doesn't return rows.
